requirements.txt
discord.py>=2.3.0
pytest>=7.4.0
